Skip over unavailable commands at the beginning
Iterators over available commands may return an unavailable command if it
is at the beginning of the commands in the current context.

The ++ and -- operators will skip other commands automatically. It doesn't
matter if the end is an unavailable command because it is not valid to
access the end of an iterator.